+++ src/main/scala/beam/playground/akkaguice/ActorInject.scala
@@ -1,4 +1,4 @@
-package beam.agentsim.akkaguice
+package beam.playground.akkaguice
 
 import akka.actor._
 import com.google.inject.Injector
